moved header components on router to root
